split update and detail product page
